extract making attachment to a contributed util
